+++ backend/src/test/java/com/zzang/chongdae/notification/domain/RoomStatusNotificationTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.google.firebase.messaging.Message;
 import com.zzang.chongdae.global.service.ServiceTest;
 import com.zzang.chongdae.member.repository.entity.MemberEntity;
+import com.zzang.chongdae.notification.domain.notification.RoomStatusNotification;
 import com.zzang.chongdae.notification.service.FcmMessageManager;
 import com.zzang.chongdae.offering.repository.entity.OfferingEntity;
 import org.junit.jupiter.api.DisplayName;
